feat: implement sharded path generation for media files and move MediaService and MediaVariantService
