+++ file/src/main/java/pt/estga/file/services/naming/FileNamingService.java
@@ -2,25 +2,40 @@ package pt.estga.file.services.naming;
 
 import org.springframework.stereotype.Service;
 import org.springframework.util.StringUtils;
+import pt.estga.file.entities.MediaFile;
 
 /**
- * Generates stored filenames for media files. Keeps naming strategy centralized for reuse.
+ * Generates stored filenames and storage paths for media files.
  */
 @Service
 public class FileNamingService {
 
-    /**
-     * Generates a stored filename based on a UUID. This avoids coupling filenames
-     * to database-generated ids and produces safe, deterministic filenames.
-     * The original filename's extension is preserved when present.
-     *
-     * @param originalFilename optional original filename used only to preserve extension
-     * @return uuid-based filename (example: 3f9e7a1c-... .jpg)
-     */
     public String generateStoredFilename(String originalFilename) {
         String extension = StringUtils.getFilenameExtension(originalFilename);
         String uuid = java.util.UUID.randomUUID().toString();
         return uuid + (extension != null ? "." + extension : "");
+    }
+
+    public String generatePath(MediaFile mediaFile) {
+        return generatePath(mediaFile.getFilename());
+    }
+
+    public String generatePath(String filename) {
+        if (filename == null) {
+            throw new IllegalArgumentException("filename cannot be null");
+        }
+
+        filename = filename.replace("\\", "/");
+
+        String base = filename;
+        int dot = filename.indexOf('.');
+        if (dot > 0) base = filename.substring(0, dot);
+
+        String part = base.length() >= 4 ? base.substring(0,4) : String.format("%4s", base).replace(' ', 'x');
+        String p1 = part.substring(0,2);
+        String p2 = part.substring(2,4);
+
+        return String.format("%s/%s/%s", p1, p2, filename);
     }
 }
